refactor: Move queue.ts and its tests
