+++ tests/dataStructures/queue.unit.test.ts
@@ -1,5 +1,4 @@
-import fc from "fast-check";
-import { createQueue } from "./queue";
+import { createQueue } from "../../src/dataStructures/queue";
 
 describe("unit testing for Queue", () => {
   test("should wait when popping an empty queue", async () => {
@@ -33,32 +32,3 @@ describe("unit testing for Queue", () => {
     expect(queue.peek()).toBeUndefined();
   });
 });
-
-/* eslint-disable no-restricted-syntax,no-await-in-loop */
-describe("property-based testing for Queue", () => {
-  test("should correctly push item into the queue", () => {
-    fc.assert(
-      fc.property(fc.anything(), (value) => {
-        const queue = createQueue();
-        queue.push(value);
-        expect(queue.size()).toBe(1);
-        expect(queue.peek()).toEqual(value);
-      }),
-    );
-  });
-
-  test("should correctly pop items in FIFO order", async () => {
-    await fc.assert(
-      fc.asyncProperty(fc.array(fc.anything()), async (values) => {
-        const queue = createQueue();
-        values.forEach((value) => {
-          queue.push(value);
-        });
-        for (const value of values) {
-          expect(await queue.pop()).toEqual(value);
-        }
-      }),
-    );
-  });
-});
-/* eslint-enable no-restricted-syntax,no-await-in-loop */
